Pull from remote before updating module repo; Handle errors when changing source in rockspec
We now try to pull all changes when updating a module.
If there is a remote specified for the repository, we fetch all
remote branches and tags. Then for each remote branch 'name'
we run 'git checkout -f -B name origin/name' which creates the
respective local branch if it doesn't already exist and then
sets its starting point to the remote branch, wich essentially
'clones' the branch to the local repository.

When updating the module source in the attached rockspec file,
if the text-based parser should fail to find and comment out the
original source, we load the rockspec file as a Lua table,
then we change source and write the table back to the file.
This makes sure the proper source is always specified.