Allow more single-letter prefixes in version strings
Some modules have letter prefixes in their version strings.
Until now, only the "v" prefix was allowed. Now we also allow
prefixes "a", "b" and "r" for "alpha", "beta" and "release/RC" versions.

Also fixed a comparison error when the major version could not be parsed.
Now when the major version cannot be determined, the module is not processed
further and an error-level message is logged.